requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-cleanup==8.0.0
django-cors-headers==4.2.0
django-crontab==0.7.1
django-filter==23.2
django-import-export==3.2.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
phonenumberslite==8.13.18
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
razorpay==1.4.1
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
